Fix some cases where we're shifting a signed 1 left.
Shift 1U instead, to make sure it's unsigned; the result of, for
example, the result of shifting a signed value left is undefined if the
value times 2^{shift count} doesn't fit in the *signed* type of the
shifted value.  That means, in particular, that the result of shifting 1
left by {number of bits in an int - 1} is undefined.  (In *practice*,
it'll probably be -2^32, with the bit you want set, but that's not
guaranteed, and GCC 5.1 seems not to like it.)

Reviewed-on: https://code.wireshark.org/review/8255

+++ epan/req_resp_hdrs.c
@@ -288,7 +288,7 @@ req_resp_hdrs_do_reassembly(tvbuff_t *tvb, const int offset, packet_info *pinfo,
 					 */
 					return TRUE;
 				}
-				if (chunk_size > (guint)1<<31) {
+				if (chunk_size > 1U<<31) {
 					/* Chunk size is unreasonable. */
 					/* XXX What /is/ reasonable? */
 					return TRUE;
